Add and consume DiagnosticsWriter class
